adds get_index, delete_file, and revamps startup.py.

+++ UploadDownload/models.py
@@ -8,16 +8,17 @@ class SingleFileUpload(models.Model):
     def loc_func(self, filename):
         return os.path.join(settings.CLOUD_DIR, self.username, self.file_path, str(self.file))
     file_path = models.CharField(max_length=100)
-    file = models.FileField(upload_to=loc_func, max_length=1000)
+    file = models.FileField(upload_to=loc_func, max_length=1000, default=None)
     md5sum = models.CharField(max_length=200)
     file_url = models.CharField(max_length=200)
     username = models.CharField(max_length=100)
+    name = models.CharField(max_length=200)
 
     class Meta:
         unique_together = (("username", "md5sum", "file_path"), )
 
     def __str__(self):
-        return self.file.name+'@'+self.user+'md5'+self.md5sum
+        return self.file.name+'@'+self.username+'md5'+self.md5sum
 
 
 @receiver(models.signals.post_delete, sender=SingleFileUpload)
@@ -26,6 +27,7 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
     Deletes file from filesystem
     when corresponding `MediaFile` object is deleted.
     """
+    print(instance.file)
     if instance.file:
         if os.path.isfile(instance.file.path):
             os.remove(instance.file.path)
@@ -46,6 +48,7 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
         return False
 
     new_file = instance.file
-    if not old_file == new_file:
-        if os.path.isfile(old_file.path):
-            os.remove(old_file.path)
+    if old_file:
+        if not old_file == new_file:
+            if os.path.isfile(old_file.path):
+                os.remove(old_file.path)
